requirements.txt
numpy>=1.21.0
numba>=0.56.0
matplotlib>=3.5.0
pillow>=9.0.0
scipy>=1.7.0
PyQt5>=5.15.0
pyqtgraph>=0.12.0
colorcet>=3.0.0